🐞 Fix race condition for Azure image creation
This fixes a failure which occurs 50% of the time (in a sample set of our
six most recent runs).

The failure is caused by a race condition precipitated by the manner in
which Azure CPI treats stemcells; specifically, the Azure CPI stemcell
image is not part of the `bosh upload-stemcell` command, but rather an
artifact of the first `bosh deploy` using that stemcell.

Furthermore, the failure _only_ occurs when a fresh stemcell is uploaded
and more than one deployment using that fresh stemcell is simultaneously
kicked off. We judge this series of events to be sufficiently rare (Ops
Manager only kicks off one deployment at a time) that we address this in
our test rather than refactoring the Azure CPI.

Sample failure:
```
Task 11721 | 17:43:48 | Creating missing vms: check-windowsdefender/f6453233-4abc-4f8f-977b-cbae181ccb47 (0) (00:00:31)
                      L Error: CPI error 'Bosh::Clouds::VMCreationFailed' with message '#<Bosh::AzureCloud::AzureConflictError: http_put - http code: 409
x-ms-client-request-id: 147fa503-99b7-4aa1-a40b-236acb7f1a09
x-ms-request-id: af41449c-b5eb-40d9-8cd2-793826645b60
x-ms-correlation-request-id: 6fda8abb-af3d-43f1-8678-242eb443308d
x-ms-routing-request-id: EASTUS:20220217T174408Z:6fda8abb-af3d-43f1-8678-242eb443308d
Error message: {
  "error": {
    "code": "OperationNotAllowed",
    "message": "The resource '2b0bfac1-59c7-469f-98ee-91b577e9fa41' cannot be created from Image '/subscriptions/305c1785-7cd3-4605-821c-5f430f9968bf/resourceGroups/azure-director/providers/Microsoft.Compute/images/630529f1-23e7-40aa-8c1c-8fec952046ec-S-eastus' until Image has been successfully created."
  }
}>
/var/vcap/data/packages/bosh_azure_cpi/35b71bc5275869f976f8ac18ffdab8b8a90f4e41/lib/cloud/azure/restapi/azure_client.rb:2442:in `check_completion'
/var/vcap/data/packages/bosh_azure_cpi/35b71bc5275869f976f8ac18ffdab8b8a90f4e41/lib/cloud/azure/restapi/azure_client.rb:2530:in `http_put'
/var/vcap/data/packages/bosh_azure_cpi/35b71bc5275869f976f8ac18ffdab8b8a90f4e41/lib/cloud/azure/restapi/azure_client.rb:444:in `create_virtual_machine'
```

Signed-off-by: Brian Upton <bupton@vmware.com>